+++ assets/js/platformer2/Audio.js
@@ -4,6 +4,24 @@ function playGoombaDeath() {
   goombaDeathSound.play();
 }
 
+//Audio when jumping
+function playJump() {
+  const PlayerJumpSound = document.getElementById("PlayerJump");
+  PlayerJumpSound.play();
+}
+
+//Audio when player Dies
+function playPlayerDeath() {
+  const playerDeathSound = document.getElementById("PlayerDeath");
+  playerDeathSound.play();
+}
+
+//Audio when stepping on Mushroom
+function mushroomSound() {
+  const mushroomSound = document.getElementById("Mushroom");
+  mushroomSound.play();
+}
+
 /*To Add Other Audio Functions:
   function soundFunction() {
     const soundVariable = document.getElementById("soundID");
@@ -13,4 +31,9 @@ function playGoombaDeath() {
 
 
 //Export Sound
-export default playGoombaDeath
+export default {
+  playGoombaDeath: () => { playGoombaDeath() },
+  playJump: () => { playJump() },
+  playPlayerDeath: () => { playPlayerDeath() },
+  mushroomSound: () => { mushroomSound() },
+}
