--- assets/js/platformer2/Audio3.js
@@ -1,10 +1,0 @@
-//Audio when stepping on Mushroom
-function MushroomSound() {
-  const mushroomSound = document.getElementById("Mushroom");
-  mushroomSound.play();
-}
-
-
-//Export Sound
-export default MushroomSound
-
