+++ assets/js/platformer2/FlyingGoomba.js
@@ -116,7 +116,7 @@ export class FlyingGoomba extends Character {
                 this.canvas.style.transformOrigin = "bottom"; // Set the transform origin to the bottom
                 this.canvas.style.transform = "scaleY(0)"; // Make the Goomba flat
                 this.speed = 0;
-                playGoombaDeath();
+                playGoombaDeath.playGoombaDeath();
 
                 setTimeout((function() {
                     GameEnv.invincible = false;
